Create the header file
