+++ Pokemon/PokemonChoice.hpp
@@ -1,0 +1,7 @@
+// Define an enum for Pokemon choices
+enum class PokemonChoice {
+    CHARMANDER = 1,
+    BULBASAUR,
+    SQUIRTLE,
+    PIKACHU // Default choice
+  };
