+++ Pokemon/PokemonType.hpp
@@ -1,0 +1,7 @@
+enum class PokemonType {
+    FIRE,
+    GRASS,
+    WATER,
+    ELECTRIC,
+    NORMAL // Added for the default constructor
+  };
