--- Pokemon/header.hpp
@@ -1,1 +1,0 @@
-}
